Handle missing required arguments, move _command_utils to _script_utils
